changes for frontend - urls, idea status

+++ src/main/java/com/tyiu/corn/model/dto/IdeaDTO.java
@@ -40,7 +40,7 @@ public class IdeaDTO {
     private Long marketValue;
     private Long originality;
     private Long technicalRealizability;
-    private Long preAssessment;
+    private double preAssessment;
     private double rating;
     private double risk;
 }
